requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pyperclip>=1.8.2
PyCryptodome>=3.19.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.5
SQLAlchemy>=2.0.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dateutil>=2.8.2